feat : MainViewController에서 DetailViewController분리
