+++ BoostClusteringMaB/BoostClusteringMaB/Scenes/DetailView/DetailViewController.swift
@@ -1,0 +1,134 @@
+//
+//  DetailViewController.swift
+//  BoostClusteringMaB
+//
+//  Created by 강민석 on 2020/12/03.
+//
+
+import UIKit
+import CoreData
+
+class DetailViewController: UIViewController {
+    
+    @IBOutlet var collectionView: UICollectionView!
+    
+    var fetchedResultsController: NSFetchedResultsController<ManagedPOI>?
+    
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        collectionView.delegate = self
+        collectionView.dataSource = self
+        initializeFetchedResultsController()
+    }
+    
+    func initializeFetchedResultsController() {
+        
+        let coreDataLayer = CoreDataLayer()
+        
+        fetchedResultsController = coreDataLayer.makeFetchResultsController(
+            southWest: LatLng(lat: 30, lng: 120),
+            northEast: LatLng(lat: 45, lng: 135)
+        )
+        
+        fetchedResultsController?.delegate = self
+        
+        do {
+            try fetchedResultsController?.performFetch()
+        } catch {
+            fatalError("Failed to initialize FetchedResultsController: \(error)")
+        }
+
+    }
+}
+extension DetailViewController: NSFetchedResultsControllerDelegate {
+    
+    func controller(_ controller: NSFetchedResultsController<NSFetchRequestResult>,
+                    didChange sectionInfo: NSFetchedResultsSectionInfo,
+                    atSectionIndex sectionIndex: Int,
+                    for type: NSFetchedResultsChangeType) {
+        switch type {
+        case .insert:
+            collectionView.insertSections(IndexSet(integer: sectionIndex))
+        case .delete:
+            collectionView.deleteSections(IndexSet(integer: sectionIndex))
+        case .move:
+            break
+        case .update:
+            break
+        @unknown default:
+            fatalError()
+        }
+    }
+    
+    func controller(_ controller: NSFetchedResultsController<NSFetchRequestResult>,
+                    didChange anObject: Any,
+                    at indexPath: IndexPath?,
+                    for type: NSFetchedResultsChangeType,
+                    newIndexPath: IndexPath?) {
+        switch type {
+        case .insert:
+            collectionView.insertItems(at: [newIndexPath ?? .init()])
+        case .delete:
+            collectionView.deleteItems(at: [indexPath ?? .init()])
+        case .update:
+            collectionView.reloadItems(at: [indexPath ?? .init()])
+        case .move:
+            collectionView.moveItem(at: indexPath ?? .init(), to: newIndexPath ?? .init())
+        @unknown default:
+            fatalError()
+        }
+    }
+    
+}
+
+extension DetailViewController: UICollectionViewDelegate, UICollectionViewDataSource, UICollectionViewDelegateFlowLayout {
+    
+    func collectionView(_ collectionView: UICollectionView,
+                        layout collectionViewLayout: UICollectionViewLayout,
+                        sizeForItemAt indexPath: IndexPath) -> CGSize {
+        CGSize(width: self.view.bounds.width - 20, height: 110)
+    }
+    
+    func collectionView(_ collectionView: UICollectionView,
+                        layout collectionViewLayout: UICollectionViewLayout,
+                        minimumLineSpacingForSectionAt section: Int) -> CGFloat {
+        2
+    }
+    
+    func collectionView(_ collectionView: UICollectionView, numberOfItemsInSection section: Int) -> Int {
+        guard let fetchedResultsController = fetchedResultsController,
+              let sections = fetchedResultsController.sections
+        else { return 0 }
+        
+        return sections[section].numberOfObjects
+    }
+    
+    func collectionView(_ collectionView: UICollectionView,
+                        cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
+        guard let cell = collectionView.dequeueReusableCell(withReuseIdentifier: "cell",
+                                                            for: indexPath)
+                as? DetailCollectionViewCell,
+              let object = fetchedResultsController?.object(at: indexPath)
+        else {
+            return UICollectionViewCell()
+        }
+        cell.configure(poi: object)
+        cell.layer.borderWidth = 1
+        cell.layer.cornerRadius = 10
+        
+        return cell
+    }
+    
+    func collectionView(_ collectionView: UICollectionView,
+                        viewForSupplementaryElementOfKind kind: String,
+                        at indexPath: IndexPath) -> UICollectionReusableView {
+        guard let header = collectionView.dequeueReusableSupplementaryView(ofKind: kind,
+                                                                           withReuseIdentifier: "header",
+                                                                           for: indexPath)
+                as? DetailCollectionReusableView
+        else { return UICollectionReusableView() }
+       // header.poiNumberLabel.text = "\(displayedData.count)개"
+        //나중에 dataSource.count로 표기
+        return header
+    }
+}
